# crates/lambdas/indexer-api/openapi.yaml
openapi: 3.0.3
info:
  title: Boundless Indexer API
  description: API for accessing staking, delegation, and Proof of Verifiable Work (PoVW) data for the Boundless protocol.
  version: 1.0.0
  contact:
    name: Boundless Development Team
servers:
  - url: /
    description: Current server
...
paths:
...
  /v1/delegations/votes/addresses:
    get:
      tags:
        - Delegations
      summary: Get Aggregate Vote Delegations
      description: Get aggregate vote delegation powers across all epochs
      operationId: getAggregateVoteDelegations
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DelegationAggregateResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/delegations/votes/epochs/{epoch}/addresses:
    get:
      tags:
        - Delegations
      summary: Get Vote Delegations by Epoch
      description: Get vote delegation powers for a specific epoch
      operationId: getVoteDelegationsByEpoch
      parameters:
        - $ref: '#/components/parameters/EpochParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DelegationByEpochResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/delegations/votes/addresses/{address}:
    get:
      tags:
        - Delegations
      summary: Get Vote Delegation History by Address
      description: Get vote delegation history for a specific delegate address
      operationId: getVoteDelegationHistoryByAddress
      parameters:
        - $ref: '#/components/parameters/AddressParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: start_epoch
          in: query
          description: Filter results from this epoch onwards
          schema:
            type: integer
            minimum: 0
        - name: end_epoch
          in: query
          description: Filter results up to this epoch
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DelegationByEpochResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/delegations/votes/epochs/{epoch}/addresses/{address}:
    get:
      tags:
        - Delegations
      summary: Get Vote Delegation by Address and Epoch
      description: Get vote delegation power for a specific delegate at a specific epoch
      operationId: getVoteDelegationByAddressAndEpoch
      parameters:
        - $ref: '#/components/parameters/EpochParam'
        - $ref: '#/components/parameters/AddressParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DelegationPowerByEpoch'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/delegations/rewards/addresses:
    get:
      tags:
        - Delegations
      summary: Get Aggregate Reward Delegations
      description: Get aggregate reward delegation powers across all epochs
      operationId: getAggregateRewardDelegations
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DelegationAggregateResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/delegations/rewards/epochs/{epoch}/addresses:
    get:
      tags:
        - Delegations
      summary: Get Reward Delegations by Epoch
      description: Get reward delegation powers for a specific epoch
      operationId: getRewardDelegationsByEpoch
      parameters:
        - $ref: '#/components/parameters/EpochParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DelegationByEpochResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/delegations/rewards/addresses/{address}:
    get:
      tags:
        - Delegations
      summary: Get Reward Delegation History by Address
      description: Get reward delegation history for a specific delegate address
      operationId: getRewardDelegationHistoryByAddress
      parameters:
        - $ref: '#/components/parameters/AddressParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: start_epoch
          in: query
          description: Filter results from this epoch onwards
          schema:
            type: integer
            minimum: 0
        - name: end_epoch
          in: query
          description: Filter results up to this epoch
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DelegationByEpochResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/delegations/rewards/epochs/{epoch}/addresses/{address}:
    get:
      tags:
        - Delegations
      summary: Get Reward Delegation by Address and Epoch
      description: Get reward delegation power for a specific delegate at a specific epoch
      operationId: getRewardDelegationByAddressAndEpoch
      parameters:
        - $ref: '#/components/parameters/EpochParam'
        - $ref: '#/components/parameters/AddressParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DelegationPowerByEpoch'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    OffsetParam:
      name: offset
      in: query
      description: Number of results to skip for pagination
      schema:
        type: integer
        minimum: 0
        default: 0
    EpochParam:
      name: epoch
      in: path
      required: true
      description: Epoch number (2-day period)
      schema:
        type: integer
        minimum: 0
    AddressParam:
      name: address
      in: path
      required: true
      description: Ethereum address (0x-prefixed, checksummed)
      schema:
        type: string
        pattern: '^0x[a-fA-F0-9]{40}$'
        example: '0x2408e37489c231f883126c87e8aadbad782a040a'
  schemas:
    Pagination:
      type: object
      required:
        - count
        - offset
        - limit
      properties:
        count:
          type: integer
          description: Number of items in this response
        offset:
          type: integer
          description: Offset used for this response
        limit:
          type: integer
          description: Maximum items per response
...
    DelegationPowerAggregate:
      type: object
      required:
        - rank
        - delegate_address
        - power
        - delegator_count
        - delegators
        - epochs_participated
      properties:
        rank:
          type: integer
          description: Leaderboard rank
        delegate_address:
          type: string
          description: Ethereum address of the delegate
        power:
          type: string
          description: Total delegation power (in wei)
        delegator_count:
          type: integer
          description: Number of delegators
        delegators:
          type: array
          items:
            type: string
          description: List of delegator addresses
        epochs_participated:
          type: integer
          description: Number of epochs participated
    DelegationPowerByEpoch:
      type: object
      required:
        - delegate_address
        - epoch
        - power
        - delegator_count
        - delegators
      properties:
        rank:
          type: integer
          description: Leaderboard rank (only present in list/leaderboard responses, absent for individual address queries)
        delegate_address:
          type: string
          description: Ethereum address of the delegate
        epoch:
          type: integer
          description: Epoch number
        power:
          type: string
          description: Delegation power in this epoch (in wei)
        delegator_count:
          type: integer
          description: Number of delegators
        delegators:
          type: array
          items:
            type: string
          description: List of delegator addresses
...
    DelegationAggregateResponse:
      type: object
      required:
        - entries
        - pagination
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/DelegationPowerAggregate'
        pagination:
          $ref: '#/components/schemas/Pagination'
    DelegationByEpochResponse:
      type: object
      required:
        - entries
        - pagination
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/DelegationPowerByEpoch'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Detailed error message
  responses:
    BadRequest:
      description: Bad Request - Invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Bad Request
            message: Invalid address format
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Not Found
            message: No data found for the specified parameters
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal Server Error
            message: An unexpected error occurred